Add build contexts to make template simple
GitHub environments are not easy to use in template. User has no way
to render shortcut commit SHA, has no way to check if the ref is a
branch or a tag, has to make up GitHub link in the template by
themselves which makes template complicated.

Providing build context such as branch name, tag name, links makes
it not that difficult to write a message template.